Find tables spike
