+++ elasticsearch/src/main/scala/com/stratio/crossdata/connector/elasticsearch/ElasticSearchMetadataBuilder.scala
@@ -1,0 +1,17 @@
+package com.stratio.crossdata.connector.elasticsearch
+
+import com.sksamuel.elastic4s.ElasticDsl._
+import com.sksamuel.elastic4s._
+import com.stratio.crossdata.connector.elasticsearch.ElasticSearchConnectionUtils._
+
+object ElasticSearchMetadataBuilder {
+
+
+  def listTypes(options: Map[String, String] ): Any ={
+
+    buildClient(options).admin.indices().getMappings(null).
+
+  null
+  }
+
+}
