+++ elasticsearch/src/main/scala/com/stratio/crossdata/connector/elasticsearch/ElasticSearchQueryProcessor.scala
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *         http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,6 +17,7 @@ package com.stratio.crossdata.connector.elasticsearch
 
 import com.sksamuel.elastic4s.ElasticDsl._
 import com.sksamuel.elastic4s._
+import com.stratio.crossdata.connector.elasticsearch.ElasticSearchConnectionUtils._
 import org.apache.spark.Logging
 import org.apache.spark.sql.catalyst.expressions.{Attribute, Literal}
 import org.apache.spark.sql.catalyst.planning.PhysicalOperation
@@ -26,9 +27,7 @@ import org.apache.spark.sql.sources.{CatalystToCrossdataAdapter, Filter => Sourc
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.{Row, sources}
 import org.elasticsearch.action.search.SearchResponse
-import org.elasticsearch.common.settings.ImmutableSettings
 import org.elasticsearch.hadoop.cfg.ConfigurationOptions._
-import com.stratio.crossdata.connector.elasticsearch.DefaultSource._
 
 object ElasticSearchQueryProcessor {
 
@@ -45,7 +44,6 @@ object ElasticSearchQueryProcessor {
 class ElasticSearchQueryProcessor(val logicalPlan: LogicalPlan, val parameters: Map[String, String], val schemaProvided: Option[StructType] = None) extends Logging {
 
 
-
   /**
    * Executes the [[LogicalPlan]]] and query the ElasticSearch database
    * @return the query result
@@ -58,16 +56,8 @@ class ElasticSearchQueryProcessor(val logicalPlan: LogicalPlan, val parameters: 
         (resource(0), resource(1))
       }
 
-      val host: String = parameters.getOrElse(ES_NODES, ES_NODES_DEFAULT)
-      val port: Int = parameters.getOrElse(ElasticNativePort, 9300).toString.toInt
-      val clusterName: String = parameters.get(ElasticCluster).get
-
       val finalQuery = buildNativeQuery(requiredColumns, filters, search in esIndex / esType)
-
-      val settings = ImmutableSettings.settingsBuilder().put("cluster.name", clusterName).build()
-      val client = ElasticClient.remote(settings, host, port)
-
-      val resp: SearchResponse = client.execute(finalQuery).await
+      val resp: SearchResponse = buildClient(parameters).execute(finalQuery).await
 
       ElasticSearchRowConverter.asRows(schemaProvided.get, resp.getHits.getHits, requiredColumns)
     }
@@ -85,20 +75,23 @@ class ElasticSearchQueryProcessor(val logicalPlan: LogicalPlan, val parameters: 
       case sources.StringStartsWith(attribute, value) => prefixQuery(attribute, value.toLowerCase)
     }
 
-    val searchFilters =  sFilters.collect {
-        case sources.EqualTo(attribute, value) => termFilter(attribute, value)
-        case sources.GreaterThan(attribute, value) => rangeFilter(attribute).gt(value)
-        case sources.GreaterThanOrEqual(attribute, value) => rangeFilter(attribute).gte(value.toString)
-        case sources.LessThan(attribute, value) => rangeFilter(attribute).lt(value)
-        case sources.LessThanOrEqual(attribute, value) => rangeFilter(attribute).lte(value.toString)
-        case sources.In(attribute, value) => termsFilter(attribute, value.toList: _*)
-        case sources.IsNotNull(attribute) => existsFilter(attribute)
-        case sources.IsNull(attribute) => missingFilter(attribute)
-      }
+    val searchFilters = sFilters.collect {
+      case sources.EqualTo(attribute, value) => termFilter(attribute, value)
+      case sources.GreaterThan(attribute, value) => rangeFilter(attribute).gt(value)
+      case sources.GreaterThanOrEqual(attribute, value) => rangeFilter(attribute).gte(value.toString)
+      case sources.LessThan(attribute, value) => rangeFilter(attribute).lt(value)
+      case sources.LessThanOrEqual(attribute, value) => rangeFilter(attribute).lte(value.toString)
+      case sources.In(attribute, value) => termsFilter(attribute, value.toList: _*)
+      case sources.IsNotNull(attribute) => existsFilter(attribute)
+      case sources.IsNull(attribute) => missingFilter(attribute)
+    }
 
     val matchQuery = query bool must(matchers)
 
-    val finalQuery = if (searchFilters.isEmpty) matchQuery else matchQuery postFilter bool { must(searchFilters) }
+    val finalQuery = if (searchFilters.isEmpty) matchQuery
+    else matchQuery postFilter bool {
+      must(searchFilters)
+    }
 
     log.debug("LogicalPlan transformed to the Elasticsearch query:" + finalQuery.toString())
     finalQuery
